+++ src/main/java/roomescape/member/infrastructure/JpaMemberRepository.java
@@ -1,7 +1,8 @@
-package roomescape.member.domain;
+package roomescape.member.infrastructure;
 
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
+import roomescape.member.domain.Member;
 
 public interface JpaMemberRepository extends JpaRepository<Member, Long> {
     boolean existsByEmail(String email);
