+++ src/main/java/roomescape/reservationTime/infrastructure/JpaReservationTimeRepository.java
@@ -1,6 +1,7 @@
-package roomescape.reservationTime.domain;
+package roomescape.reservationTime.infrastructure;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import roomescape.reservationTime.domain.ReservationTime;
 
 public interface JpaReservationTimeRepository extends JpaRepository<ReservationTime, Long> {
 }
