+++ src/test/java/roomescape/member/domain/JpaMemberRepositoryTest.java
@@ -11,6 +11,7 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
+import roomescape.member.infrastructure.JpaMemberRepository;
 
 @DataJpaTest
 public class JpaMemberRepositoryTest {
