+++ src/test/java/roomescape/reservationTime/domain/JpaReservationTimeRepositoryTest.java
@@ -11,6 +11,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
+import roomescape.reservationTime.infrastructure.JpaReservationTimeRepository;
 
 @DataJpaTest
 public class JpaReservationTimeRepositoryTest {
